Create %MIN_OS% variable for 'minimum_os_version'
The current minimum supported OS for Citrix Workspace is 10.15. Unfortunately the 'LSMinimumSystemVersion' key in the app bundle's 'Info.plist' doesn't seem to respect the 10.15 limitation, but the installer package blocks installed on anything under 10.15, so you can end up in an install loop where the current version is offered to a <10.15 machine, attempts to install, but always fails.
It may be possible to pull the minimum version for the 'Distribution' xml of the expanded flat package. there's an install check in there that looks like this 
```
    <script><![CDATA[
        function install_check() {
            if (system.compareVersions(system.version.ProductVersion, "10.15") < 0) {
                my.result.title = system.localizedString("InstallCheckFailedTitle");
                my.result.message = system.localizedString("InstallCheckFailedMessage");
                my.result.type = "Fatal";
                return false;
            }
            return true;
        }
    ]]></script>
```
This may be a fairly stable way too find the min version as version 19.6.0 looked like this:
```
    <script><![CDATA[
        function install_check() {
            if (system.compareVersions(system.version.ProductVersion, "10.11") < 0) {
                my.result.title = system.localizedString("InstallCheckFailedTitle");
                my.result.message = system.localizedString("InstallCheckFailedMessage");
                my.result.type = "Fatal";
                return false;
            }
            return true;
        }
    ]]></script>
```
By all means, feel free to scrape the Distribution xml for minimum version, but considering the relative infrequency that the OS minimum changes, I just went the path of least resistance and added a simple hard-coded variable.